pyproject.toml
[project]
name = "crypto-predictor"
version = "0.1.0"
description = "Crypto streaming services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "candles",
    "loguru>=0.7.3",
    "pydantic>=2.10.6",
    "quixstreams>=3.13.1",
    "requests>=2.32.3",
    "technical-indicators",
    "websocket-client>=1.8.0",
    "risingwave-py",
    "predictor",
    "mlflow>=3.1.1",
]

[project.optional-dependencies]
talib = [
    "ta-lib>=0.6.3",
]

[tool.uv.workspace]
members = ["services/trades", "candles", "services/candles", "services/technical_indicators", "services/predictor"]

[tool.uv.sources]
trades = { workspace = true }
candles = { workspace = true }
technical-indicators = { workspace = true }
predictor = { workspace = true }

[tool.ruff]
line-length = 88
lint.select = [
    "C",  # mccabe rules
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
lint.ignore = [
    "C901",  # max-complexity-10
    "E501",  # line-too-long
]

[tool.ruff.format]
indent-style = "space"
quote-style = "single"
